feat: break down read_standards into smaller files
